roll back to 4.27.5

+++ infra/conf/policy_test.go
@@ -27,7 +27,7 @@ func TestBufferSize(t *testing.T) {
 	}
 
 	for _, c := range cases {
-		bs := c.Input
+		bs := int32(c.Input)
 		pConf := Policy{
 			BufferSize: &bs,
 		}
